backend/src/apps/chatbot/serializers.py
from rest_framework import serializers
from .models import ChatMessage

class ChatMessageSerializer(serializers.ModelSerializer):
    username = serializers.CharField(source='user.username', read_only=True)
    
    class Meta:
        model = ChatMessage
        fields = ['id', 'message', 'message_type', 'timestamp', 'username']

class ChatRequestSerializer(serializers.Serializer):
    message = serializers.CharField(max_length=1000, required=True)
    
    def validate_message(self, value):
        if not value.strip():
            raise serializers.ValidationError("Message cannot be empty")
        return value.strip()